+++ src/devtools/components/main-content/object-store-view/ObjectStoreView.tsx
@@ -5,7 +5,6 @@ import Breadcrumbs from "@/devtools/components/main-content/header/Breadcrumbs";
 import MainContentHeader from "@/devtools/components/main-content/header/MainContentHeader";
 import MainContentBanner from "@/devtools/components/main-content/MainContentBanner";
 import MainContentContainer from "@/devtools/components/main-content/MainContentContainer";
-import StoreActionsDropdown from "@/devtools/components/main-content/object-store-view/StoreActionsDropdown";
 import { TableContextProvider } from "@/devtools/components/main-content/object-store-view/table-context";
 import TableStateWrapper from "@/devtools/components/main-content/object-store-view/TableStateWrapper";
 import { IndexedDB } from "@/devtools/utils/types";
@@ -18,11 +17,6 @@ export default function ObjectStoreView(props: { databases: IndexedDB[] }) {
       <MainContentContainer>
         <MainContentHeader>
           <Breadcrumbs databases={props.databases} />
-          <Show when={activeObjectStore()}>
-            {(activeStore) => (
-              <StoreActionsDropdown activeStore={activeStore()} />
-            )}
-          </Show>
         </MainContentHeader>
         <Show
           when={activeObjectStore()}