+++ src/devtools/components/sidebar/database-tree/clear-store-context.tsx
@@ -3,6 +3,13 @@ import { createStore } from "solid-js/store";
 
 import ClearStoreSidebarModal from "@/devtools/components/sidebar/database-tree/ClearStoreSidebarModal";
 import { useDatabaseTreeContext } from "@/devtools/components/sidebar/database-tree/database-tree-context";
+import {
+  createDataMutation,
+  Mutation,
+} from "@/devtools/utils/create-data-mutation";
+import { isDataMutationSuccessful } from "@/devtools/utils/inspected-window-data-mutation";
+import { triggerStoreClear } from "@/devtools/utils/inspected-window-store-clear";
+import { StoreClearRequest } from "@/devtools/utils/types";
 
 const ClearStoreContext = createContext<ClearStoreContextType>();
 
@@ -21,8 +28,28 @@ export function ClearStoreContextProvider(props: FlowProps) {
     store.modalRef?.showModal();
   };
 
+  // place the clear-store mutation in this provider to allow a single clear
+  // request at any time.
+  const { mutation: clearStoreMutation, mutate: clearStore } =
+    createDataMutation(async (request: StoreClearRequest) => {
+      await triggerStoreClear(request);
+      await isDataMutationSuccessful(
+        "__indexeddb_browser_store_clear",
+        request.requestID,
+        15_000,
+      );
+    });
+
   return (
-    <ClearStoreContext.Provider value={{ store, openModal, registerModalRef }}>
+    <ClearStoreContext.Provider
+      value={{
+        store,
+        openModal,
+        registerModalRef,
+        clearStoreMutation,
+        clearStore,
+      }}
+    >
       {props.children}
       <ClearStoreSidebarModal
         onClose={() => {
@@ -49,6 +76,8 @@ interface ClearStoreContextType {
   store: ClearStore;
   openModal: (dbIndex: number, storeIndex: number) => void;
   registerModalRef: (elt: HTMLDialogElement) => void;
+  clearStoreMutation: Mutation;
+  clearStore: (params: StoreClearRequest) => Promise<void>;
 }
 
 type ClearStore =